+++ config/di-providers-web.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Yiisoft\Yii\Debug\Api\Provider\DebugApiProvider;
+use Yiisoft\Yii\Debug\Api\Debug\Provider\DebugApiProvider;
 
 if (!(bool)($params['yiisoft/yii-debug-api']['enabled'] ?? false)) {
     return [];
