+++ config/routes.php
@@ -8,13 +8,13 @@ use Yiisoft\DataResponse\Middleware\FormatDataResponseAsJson;
 use Yiisoft\Router\Group;
 use Yiisoft\Router\Route;
 use Yiisoft\Validator\ValidatorInterface;
-use Yiisoft\Yii\Debug\Api\Controller\CacheController;
-use Yiisoft\Yii\Debug\Api\Controller\CommandController;
-use Yiisoft\Yii\Debug\Api\Controller\ComposerController;
-use Yiisoft\Yii\Debug\Api\Controller\DebugController;
-use Yiisoft\Yii\Debug\Api\Controller\GitController;
-use Yiisoft\Yii\Debug\Api\Controller\InspectController;
-use Yiisoft\Yii\Debug\Api\Middleware\ResponseDataWrapper;
+use Yiisoft\Yii\Debug\Api\Debug\Controller\DebugController;
+use Yiisoft\Yii\Debug\Api\Debug\Middleware\ResponseDataWrapper;
+use Yiisoft\Yii\Debug\Api\Inspector\Controller\CacheController;
+use Yiisoft\Yii\Debug\Api\Inspector\Controller\CommandController;
+use Yiisoft\Yii\Debug\Api\Inspector\Controller\ComposerController;
+use Yiisoft\Yii\Debug\Api\Inspector\Controller\GitController;
+use Yiisoft\Yii\Debug\Api\Inspector\Controller\InspectController;
 use Yiisoft\Yii\Middleware\CorsAllowAll;
 use Yiisoft\Yii\Middleware\IpFilter;
