+++ src/Debug/Exception/NotFoundException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Debug\Api\Exception;
+namespace Yiisoft\Yii\Debug\Api\Debug\Exception;
 
 final class NotFoundException extends \Exception
 {
