+++ src/Debug/ModuleFederationAssetBundle.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Debug\Api;
+namespace Yiisoft\Yii\Debug\Api\Debug;
 
 if (!class_exists(\Yiisoft\Assets\AssetBundle::class)) {
     class AssetBundle
