+++ src/Debug/ModuleFederationProviderInterface.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Debug\Api;
+namespace Yiisoft\Yii\Debug\Api\Debug;
 
+use Yiisoft\Yii\Debug\Api\ModuleFederationAssetBundle;
 use Yiisoft\Yii\Debug\Collector\CollectorInterface;
 
 interface ModuleFederationProviderInterface extends CollectorInterface
