+++ src/Inspector/Controller/CommandController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Debug\Api\Controller;
+namespace Yiisoft\Yii\Debug\Api\Inspector\Controller;
 
 use InvalidArgumentException;
 use Psr\Container\ContainerInterface;
