+++ src/Inspector/Controller/GitController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Debug\Api\Controller;
+namespace Yiisoft\Yii\Debug\Api\Inspector\Controller;
 
 use Gitonomy\Git\Commit;
 use Gitonomy\Git\Reference\Branch;
